requirements.txt
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
decorator==4.4.2
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
moviepy==1.0.3
numpy==1.24.2
oauth2client==4.1.3
Pillow==9.5.0
proglog==0.1.10
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDrive2==1.15.3
pyOpenSSL==23.1.1
pyparsing==3.0.9
pytube==12.1.3
PyYAML==6.0
requests==2.28.2
rsa==4.9
six==1.16.0
tqdm==4.65.0
uritemplate==4.1.1
urllib3==1.26.15
